Added: Image Upload feature for restaurants

+++ src/restaurants/restaurants.controller.ts
@@ -7,12 +7,15 @@ import {
   Post,
   Put,
   Query,
+  UseInterceptors,
+  UploadedFiles,
 } from '@nestjs/common';
 import { RestaurantService } from './restaurants.service';
 import { Restaurant } from './schemas/restaurants.schema';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
 import { Query as ExpressQuery } from 'src/utils/type';
+import { FilesInterceptor } from '@nestjs/platform-express';
 
 @Controller('restaurants')
 export class RestaurantController {
@@ -42,7 +45,29 @@ export class RestaurantController {
   }
 
   @Delete(':id')
-  async delete(@Param('id') id: string): Promise<Restaurant> {
-    return this.restaurantService.delete(id);
+  async delete(@Param('id') id: string): Promise<{ deleted: Boolean }> {
+    const restaurant = await this.restaurantService.findOne(id);
+
+    const isDeletedImages = await this.restaurantService.deleteImages(
+      restaurant.images,
+    );
+
+    if (isDeletedImages) {
+      this.restaurantService.delete(id);
+      return { deleted: true };
+    } else {
+      return { deleted: false };
+    }
+  }
+
+  @Put('upload/:id')
+  @UseInterceptors(FilesInterceptor('files'))
+  async uploadFiles(
+    @Param('id') id: string,
+    @UploadedFiles() files: Array<Express.Multer.File>,
+  ) {
+    await this.restaurantService.findOne(id);
+    const res = await this.restaurantService.uploadImages(id, files);
+    return res;
   }
 }
